pyproject.toml
[project]
name = "llm-code-review-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "fastmcp>=2.3.4",
    "gitpython>=3.1.44",
    "mcp[cli]>=1.9.0",
    "openai>=1.79.0",
    "openai-agents>=0.0.17",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.2",
    "ray[default]==2.41.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "nmagents>=0.1.0",
    "load-dotenv>=0.1.0",
]
